+++ src/modules/users/app/use_cases/CreateUserUseCase.ts
@@ -2,6 +2,7 @@ import { UserRepository } from '../datastore/User.repo';
 import { CreateUserDto } from '../../infra/controllers/dto/CreateUserDto';
 import { User } from '../../domain/entities/user';
 import { Inject, Injectable } from '@nestjs/common';
+import { ConflictException } from '@nestjs/common';
 
 @Injectable()
 export class CreateUserUseCase {
@@ -11,6 +12,18 @@ export class CreateUserUseCase {
   ) {}
 
   async handle(userDto: CreateUserDto): Promise<User> {
+    let foundUser: User | null = await this.userRepository.getUserByEmail(
+      userDto.email,
+    );
+    if (foundUser) {
+      throw new ConflictException('Email already in use');
+    }
+    foundUser = await this.userRepository.getUserByPhoneNumber(
+      userDto.phoneNumber,
+    );
+    if (foundUser) {
+      throw new ConflictException('Phone number already in use');
+    }
     const userToSave = new User(
       crypto.randomUUID(),
       userDto.name,