+++ src/modules/users/domain/entities/seller.ts
@@ -1,3 +1,5 @@
+import { AccountStatus } from '../enums/AccountStatus';
+
 export class Seller {
   constructor(
     public id: string,
@@ -8,5 +10,6 @@ export class Seller {
     public address: string,
     public phoneNumber: string,
     public email: string,
+    public accountStatus: AccountStatus,
   ) {}
 }