+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Technical
 - [AMP-4](https://mickedev.atlassian.net/browse/AMP-4) APIs to CRUD of buyer users
+- [AMP-4](https://mickedev.atlassian.net/browse/AMP-4) APIs to CRUD of buyer users
+- [AMP-5](https://mickedev.atlassian.net/browse/AMP-5) APIs to CRUD of seller users
 
 ## Fixes